finished refactoring everything
