+++ src/bro/index.js
@@ -1,0 +1,3 @@
+const Bro = require('./api')
+
+module.exports = Bro
