+++ src/browser/index.js
@@ -1,0 +1,7 @@
+const BrowserHelper = require('./api')
+const listeners = require('./listeners')
+
+module.exports = {
+  BrowserHelper,
+  listeners
+}
