+++ src/ipfs/index.js
@@ -1,0 +1,3 @@
+const IPFSHelper = require('./api')
+
+module.exports = IPFSHelper
